Move URLSplit interface definition to model
