+++ src/app/config/config.service.ts
@@ -3,17 +3,12 @@ import { Injectable, NgZone } from '@angular/core';
 import _ from 'lodash-es';
 
 import { ElectronService } from '../electron.service';
-import { NotificationType } from '../model';
+import { NotificationType, URLSplit } from '../model';
 import { NotificationService } from '../notification/notification.service';
 import { ConfigSchema as Config, CustomAction } from './config.model';
 
 interface HttpHeader {
   headers: HttpHeaders;
-}
-
-export interface URLSplit {
-  host: string;
-  port: number;
 }
 
 @Injectable({
