+++ src/locale/messages.xlf
@@ -52,7 +52,7 @@
         <source>Please restart your system. If the issue persists open an issue on GitHub.</source>
         <context-group purpose="location">
           <context context-type="sourcefile">src/app/config/config.service.ts</context>
-          <context context-type="linenumber">41</context>
+          <context context-type="linenumber">36</context>
         </context-group>
       </trans-unit>
       <trans-unit id="invalid-config" datatype="html">
@@ -123,7 +123,7 @@
         <source>no instances found.</source>
         <context-group purpose="location">
           <context context-type="sourcefile">src/app/config/setup/discover-octoprint/discover-octoprint.component.ts</context>
-          <context context-type="linenumber">54</context>
+          <context context-type="linenumber">55</context>
         </context-group>
       </trans-unit>
       <trans-unit id="prompt-extruder-info" datatype="html">
